MediaServer holder
