+++ src/main/java/net/pms/network/mediaserver/UPNPControl.java
@@ -1,3 +1,22 @@
+/*
+ * Universal Media Server, for streaming any media to DLNA
+ * compatible renderers based on the http://www.ps3mediaserver.org.
+ * Copyright (C) 2012 UMS developers.
+ *
+ * This program is a free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; version 2
+ * of the License only.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
 package net.pms.network.mediaserver;
 
 import java.awt.event.ActionEvent;
@@ -832,7 +851,7 @@ public class UPNPControl {
 									String value = entry.getValue().toString();
 									if (isNotBlank(value)) {
 										sb.append("\n").append(entry.getKey()).append(":\n  ");
-										sb.append(value.replaceAll(",", "\n  "));
+										sb.append(value.replace(",", "\n  "));
 									}
 								}
 							}
